requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-dotenv>=1.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
werkzeug>=3.0.0
google-generativeai>=0.8.0
eventlet>=0.33.0
cryptography>=41.0.0
pyOpenSSL>=24.0.0